+++ src/deployment/exceptions/LeviLaminaPluginNotFoundError.ts
@@ -1,0 +1,5 @@
+export class LeviLaminaPluginNotFoundError extends Error {
+    public constructor(pluginName: string) {
+        super(`LeviLamina plugin ${pluginName} does not found.`);
+    }
+}