+++ src/packager/exceptions/ManifestFileNotFoundError.ts
@@ -1,11 +1,11 @@
-import { CliLoggableError } from "../cli/CliLoggableError";
+import { CliLoggableError } from "../../cli/exceptions/CliLoggableError";
 
 
-export class NodeJsConfigurationFileNotFoundError extends Error implements CliLoggableError {
+export class ManifestFileNotFoundError extends Error implements CliLoggableError{
     private readonly msg: string;
 
     public constructor(fileDirectory: string) {
-        const message: string = `Could not find package.json in ${fileDirectory}.`;
+        const message: string = `Could not find manifest.json in ${fileDirectory}.`;
 
         super(message);
 
@@ -18,7 +18,7 @@ export class NodeJsConfigurationFileNotFoundError extends Error implements CliLo
 
     public getSuggestion(): Array<string> {
         const suggestion: Array<string> = new Array<string>();
-        suggestion.push("Try `npm init` to initialize the project.");
+        suggestion.push("Try `npx lses manifest` to manifest the project before.");
 
         return suggestion;
     }