+++ src/packager/exceptions/ProjectNotManifestedError.ts
@@ -1,4 +1,4 @@
-import { CliLoggableError } from "../cli/CliLoggableError";
+import { CliLoggableError } from "../../cli/exceptions/CliLoggableError";
 
 
 export class ProjectNotManifestedError extends Error implements CliLoggableError {
@@ -15,7 +15,7 @@ export class ProjectNotManifestedError extends Error implements CliLoggableError
     getSuggestion(): Array<string> {
         const suggestion: Array<string> = new Array<string>();
 
-        suggestion.push("Try `npx lses manifest` to manifest the project.");
+        suggestion.push("Try `npx lses manifest` to manifest the project before.");
 
         return suggestion;
     }